Added OptionalRedisCacheStore to allow connection failures between redis cache to be silently logged and not disrupt the caller.
