+++ cache_store_redis/.rspec
@@ -1,3 +1,4 @@
+require spec_helper
 --color
 --require spec_helper
 --format doc
