+++ cache_store_redis/lib/cache_store_redis/version.rb
@@ -1,3 +1,3 @@
 module CacheStoreRedis
-  VERSION = '0.5.5'
+  VERSION = '0.6.0'
 end
