+++ cache_store_redis/spec/test_object.rb
@@ -1,0 +1,4 @@
+class TestObject
+  attr_accessor :text
+  attr_accessor :numeric
+end
